Move filetype plugins to start
Packer doesn't reset the filetype after loading these, i.e. when editing
Cargo.toml, packer loads the TOML plugin, but then I need to explicitly
set the filetype to TOML again for the syntax to show up.

I could fix this by adding a config function that finds all open buffers
with a given filetype, sets the filetype to "", and then back to the
given value, like this:

```lua
use({
  "cespare/vim-toml",
  ft = "toml",
  config = function()
    for _, b in ipairs(vim.tbl_filter(function(b)
      return vim.bo[b].filetype == "toml"
    end, vim.api.nvim_list_bufs())) do
      vim.bo[b].filetype = ""
      vim.bo[b].filetype = "toml"
    end
  end,
})
```

But that's pretty hacky for not much benefit - it doesn't really make an
appreciable difference whether these small plugins are always loaded.