Fix packer compile autocmd
This fixes a fun new problem after converting to using
`nvim_create_autocmd()` calls. Because I have an autocmd that sources
config files in my `init.lua`, when saving `packer.lua`, it gets sourced
by that autocmd. That causes the `nvim_create_augroup()` to be called
which clears the autocmd that does the Packer compile, so it then does
_not_ get called to compile Packer. It then gets recreated.

Essentially, before this change, the autocmd was created, cleared before
it could run, then created again, cleared before it could run again,
etc.