Remove Alacritty config, move to WezTerm full time
Alacritty breaks stuff on macOS with seemingly every update and the
general response to problems is "none of the contributors use macOS."
Version 0.11.0 broke the behavior of the Option key which had worked
(with config hacks) for years, but the response to the bug report was
that its working was a coincidence and the breakage is not something
they're going to consider fixing.

WezTerm is configured with Lua (neat!) and the Option key still works.
It has too many features for my tastes but it's relatively easy to turn
them all off. Seems like a nice alternative for now.