Ensure test runner refresh runs the correct test
This fixes the situation where we run one test, then a second one, then
try to re-run the second test, but the refresh function runs the first
test again. Instead of storing the test command in the refresh function,
we're now assigning it to a buffer variable, so that the refresh
function always runs whatever test was run last.