Fix autoclose test
Using Lua as the filetype now reports the node type for the entire line
of this test as "chunk" (thus never showing that any part of the test
line is in a string) and I'm unable to determine why. Switching the
filetype to JavaScript makes it work and that's good enough.