Restore Fugitive `<leader>gw` keymap
I moved to using Gitsigns for this keymap, however, that doesn't work if
Gitsigns has not attached to the buffer. Restore the Fugitive map so I
can write buffers that aren't tracked.

This also adds a `!` to the map so it adds ignored files.

Reference commit 41ea2a11418631b51a48c2efba8709d5e6f09565